ct: Workaround Sierra clipboard issues.
For some reason, processes under tmux no longer have access to the
clipboard.  To workaround this, we need to wrap any command that
accesses the clipboard with the command, `reattach-to-user-namespace`
(which can be installed via `brew`).

A quick and easy way to do this is to wrap the shell tmux creates in
each window.