extend cd so ... changes to the project directory
Lately, I've been working in large repositories with very deep
nesting.

After changing to a deeply nested directory wihtin a project
it's annoying changing back to the project root directory. This adjusts
my `cd` function to change to the project root when `...` is the dir.

Note that because I'm using `cd` to move up the directory tree,
care is taken to ensure the search is run in a subshell.