make lazygit open lazygit in another window
i get why a lot of vim users open things in a split window: when a terminal-ed command
terminates, its buffer will go away, but also nvim will attempt to close its window.
only if there are no more windows will it replace the current window's buffer with another buffer.

this results in a weird layout if the only other window open is nvim-tree. to smooth this out,
just open lazygit in a split window using an enhanced version of Terminal that actually allows
a command to be specified.